Add files via upload
In this project, I developed a dynamic map of the United States to visually represent and cluster data points, ensuring that each cluster had data points no more than 8 miles apart. Utilizing Python, I employed geospatial analysis techniques and libraries such as Geopandas and Folium to manipulate and display geospatial data accurately. The core of the project involved creating an algorithm to form clusters of nearby data points based on their geographical proximity. Each cluster's maximum radius was strictly maintained at 8 miles to meet the specified criteria.

The dynamic map output allows for interactive exploration, enabling users to zoom in and out to view individual clusters and data points. This visual representation not only provides a clear overview of the geographical distribution of the data but also highlights areas with higher densities of data points. This tool is invaluable for spatial analysis, revealing patterns and insights that can inform decision-making in various domains, from urban planning to environmental studies. The project showcases the power of Python in handling and visualizing complex geospatial data sets in an intuitive and accessible manner.